Move new bindings into its own cc file
