+++ bindings/pydrake/geometry/geometry_py.h
@@ -11,6 +11,9 @@
 namespace drake {
 namespace pydrake {
 
+/** Define the bounding box functions. See geometry_py_bounding_box.cc. */
+void DefineGeometryBoundingBox(py::module m);
+
 /** Defines the common elements in the drake::geometry namespace. See
  geometry_py_common.cc. */
 void DefineGeometryCommon(py::module m);
