+++ bindings/pydrake/geometry/geometry_py_common.cc
@@ -22,10 +22,6 @@
 #include "drake/geometry/geometry_version.h"
 #include "drake/geometry/in_memory_mesh.h"
 #include "drake/geometry/mesh_source.h"
-#include "drake/geometry/proximity/aabb.h"
-#include "drake/geometry/proximity/obb.h"
-#include "drake/geometry/proximity/triangle_surface_mesh.h"
-#include "drake/geometry/proximity/volume_mesh.h"
 #include "drake/geometry/proximity_properties.h"
 #include "drake/geometry/shape_specification.h"
 
@@ -658,132 +654,6 @@ void DefineShapes(py::module m) {
   }
 }
 
-void DefineBoundingBoxes(py::module m) {
-  // Define Aabb class first.
-  py::class_<Aabb> aabb_cls(m, "Aabb", doc.Aabb.doc);
-  aabb_cls  // BR
-      .def(py::init<const Vector3<double>&, const Vector3<double>&>(),
-          py::arg("p_HoBo"), py::arg("half_width"), doc.Aabb.ctor.doc)
-      .def("center", &Aabb::center, py_rvp::reference_internal,
-          doc.Aabb.center.doc)
-      .def("half_width", &Aabb::half_width, py_rvp::reference_internal,
-          doc.Aabb.half_width.doc)
-      .def("lower", &Aabb::lower, doc.Aabb.lower.doc)
-      .def("upper", &Aabb::upper, doc.Aabb.upper.doc)
-      .def("pose", &Aabb::pose, doc.Aabb.pose.doc)
-      .def("CalcVolume", &Aabb::CalcVolume, doc.Aabb.CalcVolume.doc)
-      .def("Equal", &Aabb::Equal, py::arg("other"), doc.Aabb.Equal.doc)
-      .def(py::pickle(
-          [](const Aabb& self) {
-            return std::make_pair(self.center(), self.half_width());
-          },
-          [](std::pair<Vector3<double>, Vector3<double>> data) {
-            return Aabb(data.first, data.second);
-          }));
-  DefCopyAndDeepCopy(&aabb_cls);
-
-  // Define Obb class.
-  py::class_<Obb> obb_cls(m, "Obb", doc.Obb.doc);
-  obb_cls  // BR
-      .def(py::init<const math::RigidTransformd&, const Vector3<double>&>(),
-          py::arg("X_HB"), py::arg("half_width"), doc.Obb.ctor.doc)
-      .def("center", &Obb::center, py_rvp::reference_internal,
-          doc.Obb.center.doc)
-      .def("half_width", &Obb::half_width, py_rvp::reference_internal,
-          doc.Obb.half_width.doc)
-      .def("pose", &Obb::pose, py_rvp::reference_internal, doc.Obb.pose.doc)
-      .def("CalcVolume", &Obb::CalcVolume, doc.Obb.CalcVolume.doc)
-      .def("Equal", &Obb::Equal, py::arg("other"), doc.Obb.Equal.doc)
-      .def(py::pickle(
-          [](const Obb& self) {
-            return std::make_pair(self.pose(), self.half_width());
-          },
-          [](std::pair<math::RigidTransformd, Vector3<double>> data) {
-            return Obb(data.first, data.second);
-          }));
-  DefCopyAndDeepCopy(&obb_cls);
-
-  // Now add cross-referencing HasOverlap static methods.
-  // Aabb static methods.
-  aabb_cls.def_static("HasOverlap",
-      py::overload_cast<const Aabb&, const Aabb&, const math::RigidTransformd&>(
-          &Aabb::HasOverlap),
-      py::arg("a_G"), py::arg("b_H"), py::arg("X_GH"),
-      doc.Aabb.HasOverlap.doc_aabb_aabb);
-
-  aabb_cls.def_static("HasOverlap",
-      py::overload_cast<const Aabb&, const Obb&, const math::RigidTransformd&>(
-          &Aabb::HasOverlap),
-      py::arg("aabb_G"), py::arg("obb_H"), py::arg("X_GH"),
-      doc.Aabb.HasOverlap.doc_aabb_obb);
-
-  // Obb static methods.
-  obb_cls.def_static("HasOverlap",
-      py::overload_cast<const Obb&, const Obb&, const math::RigidTransformd&>(
-          &Obb::HasOverlap),
-      py::arg("a_G"), py::arg("b_H"), py::arg("X_GH"),
-      doc.Obb.HasOverlap.doc_obb_obb);
-
-  obb_cls.def_static("HasOverlap",
-      py::overload_cast<const Obb&, const Aabb&, const math::RigidTransformd&>(
-          &Obb::HasOverlap),
-      py::arg("obb_G"), py::arg("aabb_H"), py::arg("X_GH"),
-      doc.Obb.HasOverlap.doc_obb_aabb);
-
-  obb_cls.def_static("HasOverlap",
-      py::overload_cast<const Obb&, const HalfSpace&,
-          const math::RigidTransformd&>(&Obb::HasOverlap),
-      py::arg("bv_H"), py::arg("hs_C"), py::arg("X_CH"),
-      doc.Obb.HasOverlap.doc_obb_halfspace);
-
-  // AabbMaker and ObbMaker utility functions
-  // Instead of binding the classes directly (which have lifetime issues with
-  // the std::set<int>& parameter), we provide simple functions that do the
-  // complete workflow: construct, compute, and return the result.
-
-  m.def(
-      "ComputeAabbForTriangleMesh",
-      [](const TriangleSurfaceMesh<double>& mesh_M,
-          const std::set<int>& vertices) {
-        AabbMaker<TriangleSurfaceMesh<double>> maker(mesh_M, vertices);
-        return maker.Compute();
-      },
-      py::arg("mesh_M"), py::arg("vertices"),
-      "Computes an axis-aligned bounding box for the specified vertices of a "
-      "TriangleSurfaceMesh.");
-
-  m.def(
-      "ComputeAabbForVolumeMesh",
-      [](const VolumeMesh<double>& mesh_M, const std::set<int>& vertices) {
-        AabbMaker<VolumeMesh<double>> maker(mesh_M, vertices);
-        return maker.Compute();
-      },
-      py::arg("mesh_M"), py::arg("vertices"),
-      "Computes an axis-aligned bounding box for the specified vertices of a "
-      "VolumeMesh.");
-
-  m.def(
-      "ComputeObbForTriangleMesh",
-      [](const TriangleSurfaceMesh<double>& mesh_M,
-          const std::set<int>& vertices) {
-        ObbMaker<TriangleSurfaceMesh<double>> maker(mesh_M, vertices);
-        return maker.Compute();
-      },
-      py::arg("mesh_M"), py::arg("vertices"),
-      "Computes an oriented bounding box for the specified vertices of a "
-      "TriangleSurfaceMesh.");
-
-  m.def(
-      "ComputeObbForVolumeMesh",
-      [](const VolumeMesh<double>& mesh_M, const std::set<int>& vertices) {
-        ObbMaker<VolumeMesh<double>> maker(mesh_M, vertices);
-        return maker.Compute();
-      },
-      py::arg("mesh_M"), py::arg("vertices"),
-      "Computes an oriented bounding box for the specified vertices of a "
-      "VolumeMesh.");
-}
-
 void DefineMiscFunctions(py::module m) {
   m.def("CalcVolume", &CalcVolume, py::arg("shape"), doc.CalcVolume.doc);
 
@@ -877,7 +747,6 @@ void DefineGeometryCommon(py::module m) {
   DefineInMemoryMesh(m);
   DefineMeshSource(m);
   DefineShapes(m);
-  DefineBoundingBoxes(m);
   DefineGeometrySet(m);
   DefineCollisionFilterScope(m);
   DefineCollisionFilterDeclaration(m);
